+++ component/phlare/scrape/scrape_test.go
@@ -191,6 +191,7 @@ func TestUpdateWhileScraping(t *testing.T) {
 		Logger:        util.TestFlowLogger(t),
 		Registerer:    prometheus.NewRegistry(),
 		OnStateChange: func(e component.Exports) {},
+		Clusterer:     &cluster.Clusterer{Node: cluster.NewLocalNode("")},
 	}, args)
 	require.NoError(t, err)
 	scraping := atomic.NewBool(false)